+++ src/Fluentify.Tests/Model/NestingTests/WhenEqualityIsCheckedByEquals.cs
@@ -1,0 +1,15 @@
+﻿namespace Fluentify.Model.NestingTests;
+
+public sealed class WhenEqualityIsCheckedByEquals
+    : WhenEqualityIsChecked
+{
+    private protected override bool AreEqual(Nesting? instance1, Nesting? instance2)
+    {
+        if (instance1 is not null)
+        {
+            return instance1.Equals((object?)instance2);
+        }
+
+        return true;
+    }
+}