+++ src/Fluentify.Tests/Semantics/Classes.cs
@@ -19,6 +19,7 @@ internal sealed class Classes
             GlobalContent,
             InvalidDescriptorContent,
             MultipleGenericsContent,
+            NestedContent,
             OneOfThreeIgnoredContent,
             SelfDescriptorOnIgnoredContent,
             SelfDescriptorOnOptionalContent,