+++ src/Fluentify.Tests/Snippets/Classes.Nested.cs
@@ -1,0 +1,182 @@
+﻿namespace Fluentify.Snippets;
+
+public static partial class Classes
+{
+    public const string NestedContent = """
+        namespace Fluentify.Classes.Testing
+        {
+            using System.Collections.Generic;
+
+            public sealed class Outter
+            {
+                [Fluentify]
+                public sealed class Nested
+                {
+                    public int Age { get; set; }
+
+                    public string Name { get; set; }
+
+                    public IReadOnlyList<object> Attributes { get; set; }
+                }
+            }
+        }
+        """;
+
+    public static readonly Declared Nested;
+
+    public static readonly Generated NestedWithAgeExtensions = new(
+        NestedWithAgeExtensionsContent,
+        typeof(ClassGenerator),
+        "Fluentify.Classes.Testing.Outter.NestedExtensions.WithAge");
+
+    public static readonly Generated NestedWithAttributesExtensions = new(
+        NestedWithAttributesExtensionsContent,
+        typeof(ClassGenerator),
+        "Fluentify.Classes.Testing.Outter.NestedExtensions.WithAttributes");
+
+    public static readonly Generated NestedWithNameExtensions = new(
+        NestedWithNameExtensionsContent,
+        typeof(ClassGenerator),
+        "Fluentify.Classes.Testing.Outter.NestedExtensions.WithName");
+
+    private const string NestedWithAgeExtensionsContent = """
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable enable
+        #endif
+        
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Classes.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Classes.Testing.Outter.Nested WithAge(
+                    this global::Fluentify.Classes.Testing.Outter.Nested subject,
+                    int value)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    return new global::Fluentify.Classes.Testing.Outter.Nested
+                    {
+                        Age = value,
+                        Name = subject.Name,
+                        Attributes = subject.Attributes,
+                    };
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable restore
+        #endif
+        """;
+
+    private const string NestedWithAttributesExtensionsContent = """
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable enable
+        #endif
+        
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Classes.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Classes.Testing.Outter.Nested WithAttributes(
+                    this global::Fluentify.Classes.Testing.Outter.Nested subject,
+                    params object[] values)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    global::System.Collections.Generic.IReadOnlyList<object> value = values;
+
+                    if (subject.Attributes != null)
+                    {
+                        value = subject.Attributes
+                            .Union(values)
+                            .ToArray();
+                    }
+
+                    return new global::Fluentify.Classes.Testing.Outter.Nested
+                    {
+                        Age = subject.Age,
+                        Name = subject.Name,
+                        Attributes = value,
+                    };
+                }
+
+                public static global::Fluentify.Classes.Testing.Outter.Nested WithAttributes(
+                    this global::Fluentify.Classes.Testing.Outter.Nested subject,
+                    Func<object, object> builder)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    builder.ThrowIfNull("builder");
+
+                    var instance = new object();
+
+                    instance = builder(instance);
+
+                    return subject.WithAttributes(instance);
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable restore
+        #endif
+        """;
+
+    private const string NestedWithNameExtensionsContent = """
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable enable
+        #endif
+        
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Classes.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Classes.Testing.Outter.Nested WithName(
+                    this global::Fluentify.Classes.Testing.Outter.Nested subject,
+                    string value)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    return new global::Fluentify.Classes.Testing.Outter.Nested
+                    {
+                        Age = subject.Age,
+                        Name = value,
+                        Attributes = subject.Attributes,
+                    };
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        
+        #if NET5_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+        #nullable restore
+        #endif
+        """;
+}