+++ src/Fluentify.Tests/Snippets/Records.Nested.cs
@@ -1,0 +1,187 @@
+﻿namespace Fluentify.Snippets;
+
+public static partial class Records
+{
+    public const string NestedContent = """
+        namespace Fluentify.Records.Testing
+        {
+            using System.Collections.Generic;
+
+            public sealed partial class Outter
+            {
+                [Fluentify]
+                public sealed partial record Nested(int Age, string Name, IReadOnlyList<object>? Attributes = default);
+            }
+        }
+        """;
+
+    public static readonly Declared Nested;
+
+    public static readonly Generated NestedConstructor = new(
+        NestedConstructorContent,
+        typeof(RecordGenerator),
+        "Fluentify.Records.Testing.Outter.Nested.ctor");
+
+    public static readonly Generated NestedWithAgeExtensions = new(
+        NestedWithAgeExtensionsContent,
+        typeof(RecordGenerator),
+        "Fluentify.Records.Testing.Outter.NestedExtensions.WithAge");
+
+    public static readonly Generated NestedWithAttributesExtensions = new(
+        NestedWithAttributesExtensionsContent,
+        typeof(RecordGenerator),
+        "Fluentify.Records.Testing.Outter.NestedExtensions.WithAttributes");
+
+    public static readonly Generated NestedWithNameExtensions = new(
+        NestedWithNameExtensionsContent,
+        typeof(RecordGenerator),
+        "Fluentify.Records.Testing.Outter.NestedExtensions.WithName");
+
+    private const string NestedConstructorContent = """
+        #nullable enable
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Records.Testing
+        {
+            using System.Diagnostics.CodeAnalysis;
+
+            partial class Outter
+            {
+                partial record Nested
+                {
+                    #pragma warning disable CS8604
+
+                    #if NET7_0_OR_GREATER
+                    [SetsRequiredMembers]
+                    #endif
+                    public Nested()
+                        : this(default(int), default(string), default(global::System.Collections.Generic.IReadOnlyList<object>?))
+                    {
+                    }
+
+                    #pragma warning restore CS8604
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        #nullable restore
+        """;
+
+    private const string NestedWithAgeExtensionsContent = """
+        #nullable enable
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Records.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Records.Testing.Outter.Nested WithAge(
+                    this global::Fluentify.Records.Testing.Outter.Nested subject,
+                    int value)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    return subject with
+                    {
+                        Age = value,
+                    };
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        #nullable restore
+        """;
+
+    private const string NestedWithAttributesExtensionsContent = """
+        #nullable enable
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Records.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Records.Testing.Outter.Nested WithAttributes(
+                    this global::Fluentify.Records.Testing.Outter.Nested subject,
+                    params object[] values)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    global::System.Collections.Generic.IReadOnlyList<object>? value = values;
+
+                    if (subject.Attributes != null)
+                    {
+                        value = subject.Attributes
+                            .Union(values)
+                            .ToArray();
+                    }
+
+                    return subject with
+                    {
+                        Attributes = value,
+                    };
+                }
+
+                public static global::Fluentify.Records.Testing.Outter.Nested WithAttributes(
+                    this global::Fluentify.Records.Testing.Outter.Nested subject,
+                    Func<object, object> builder)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    builder.ThrowIfNull("builder");
+
+                    var instance = new object();
+
+                    instance = builder(instance);
+
+                    return subject.WithAttributes(instance);
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        #nullable restore
+        """;
+
+    private const string NestedWithNameExtensionsContent = """
+        #nullable enable
+        #pragma warning disable CS8625
+
+        namespace Fluentify.Records.Testing
+        {
+            using System;
+            using System.Collections.Generic;
+            using System.Linq;
+            using Fluentify.Internal;
+
+            public static partial class NestedExtensions
+            {
+                public static global::Fluentify.Records.Testing.Outter.Nested WithName(
+                    this global::Fluentify.Records.Testing.Outter.Nested subject,
+                    string value)
+                {
+                    subject.ThrowIfNull("subject");
+
+                    return subject with
+                    {
+                        Name = value,
+                    };
+                }
+            }
+        }
+
+        #pragma warning restore CS8625
+        #nullable restore
+        """;
+}